Handle blocklist on message update as well
